refactor(ui): share one scrolling row between Explore and the desktop
The desktop's featured row was a second copy of the same drag-to-scroll
pattern, so the affordances added to Explore's carousel were invisible
there — which is where they were noticed missing.

Both now use one DragScrollRow: arrows in whichever direction has more to
show, a fade at those edges, and scroll snapping.

The fade is a CSS mask rather than a gradient painted in the background
colour. A coloured overlay has to know what it sits on, and the desktop's
background is not Explore's; a mask fades the content's own alpha and is
correct on any surface.
